feat: 표절 검사 실행 커맨드, HTML 변경
